requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
asgiref
async-timeout
attrs
azure-cognitiveservices-speech==1.42.0
azure-core==1.35.1
azure-identity==1.25.0
azure-storage-blob==12.26.0
Brotli
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
colorama
cryptography==43.0.0
dataclasses-json
deepl==1.22.0
defusedxml==0.8.0rc2
distro
Django
django-cors-headers
django-debug-toolbar
django-filter==24.2
django-templated-mail==1.1.1
djangorestframework
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-nested-routers==0.94.1
et_xmlfile
exceptiongroup
frozenlist
greenlet
h11
h2
hpack
httpcore
httpx
httpx-sse
hyperframe
idna
isodate==0.7.2
jiter
jsonpatch
jsonpointer
marshmallow
msal==1.34.0
msal-extensions==1.3.1
msgpack
multidict
mypy_extensions
numpy
oauthlib==3.2.2
openai==1.91.0
openpyxl
orjson
ormsgpack
packaging
pandas
propcache
psycopg2-binary==2.9.11
pycparser
pydantic
pydantic-settings
pydantic_core
PyJWT==2.9.0
PySocks
python-dateutil
python-dotenv
python3-openid==3.2.0
pytz
PyYAML
regex
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt
setuptools==70.1.1
six
sniffio
social-auth-app-django==5.4.2
social-auth-core==4.5.4
SQLAlchemy
sqlparse
tenacity
tiktoken
tqdm
typing-inspection
typing_extensions
typing_inspect
tzdata
urllib3==2.2.2
wheel==0.43.0
whitenoise==6.11.0
win_inet_pton
xxhash
yarl
zstandard==0.23.0
